requirements.txt
aicns-univariate-analyzer==0.8.3
certifi==2022.9.24
joblib==1.2.0
numpy==1.21.6
packaging==21.3
pandas==1.1.5
patsy==0.5.3
pendulum==2.1.2
pip==22.2.2
plotly==5.10.0
py4j==0.10.9.5
pyparsing==3.0.9
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2022.6
pytzdata==2020.1
scikit-learn==0.24.2
scipy==1.7.3
setuptools==65.5.0
six==1.16.0
statsmodels==0.13.2
tenacity==8.1.0
threadpoolctl==3.1.0
wheel==0.37.1
